requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-generativeai
python-dotenv
google-cloud-storage
google-genai
authlib
requests
google-cloud-secret-manager
uvicorn
python-multipart
itsdangerous
google-api-python-client
google-auth
google-auth-httplib2
